requirements.txt
aioredis==1.3.1
asgiref==3.2.9
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
google-api-core==1.20.1
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
hiredis==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
oauthlib==3.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
utils==1.0.1
zope.interface==5.1.0